Move network trace to request interceptor
